Move to own folder, update styles
