+++ ui/components/app/app-components.scss
@@ -9,6 +9,7 @@
 @import 'alerts/alerts';
 @import 'app-header/index';
 @import 'asset-list-item/asset-list-item';
+@import 'beta-header/index';
 @import 'cancel-speedup-popover/index';
 @import 'confirm-page-container/index';
 @import 'confirm-page-container/enableEIP1559V2-notice';
