+++ ui/components/app/beta-header/index.scss
@@ -1,0 +1,21 @@
+.beta-header {
+  background: var(--color-warning-default);
+  padding: 8px 8px;
+  display: flex;
+  width: 100%;
+
+  &__message {
+    text-align: center;
+    flex-grow: 1;
+  }
+
+  &__button {
+    background: transparent;
+    padding: 0 6px;
+    margin: 0;
+
+    i {
+      color: var(--color-warning-inverse);
+    }
+  }
+}
